Allow "regular setup" to save as key values
This will allow special characters to be used and easily parsed back into the UI fields
